+++ common/src/main/java/net/bettercombat/client/animation/AttackAnimationSubStack.java
@@ -3,13 +3,17 @@ package net.bettercombat.client.animation;
 import dev.kosmx.playerAnim.api.layered.ModifierLayer;
 import dev.kosmx.playerAnim.api.layered.modifier.AbstractModifier;
 import dev.kosmx.playerAnim.api.layered.modifier.MirrorModifier;
+import net.bettercombat.client.animation.modifier.AdjustmentModifier;
+import net.bettercombat.client.animation.modifier.TransmissionSpeedModifier;
 
 public class AttackAnimationSubStack {
     public final TransmissionSpeedModifier speed = new TransmissionSpeedModifier();
     public final MirrorModifier mirror = new MirrorModifier();
     public final ModifierLayer base = new ModifierLayer(null);
+    public final AdjustmentModifier adjustmentModifier;
 
-    public AttackAnimationSubStack(AbstractModifier adjustmentModifier) {
+    public AttackAnimationSubStack(AdjustmentModifier adjustmentModifier) {
+        this.adjustmentModifier = adjustmentModifier;
         mirror.setEnabled(false);
         base.addModifier(adjustmentModifier, 0);
         base.addModifier(speed, 0);