+++ common/src/main/java/net/bettercombat/client/animation/FirstPersonAnimator.java
@@ -1,7 +1,0 @@
-package net.bettercombat.client.animation;
-
-import dev.kosmx.playerAnim.api.layered.ModifierLayer;
-
-public interface FirstPersonAnimator {
-    void addLayer(ModifierLayer layer);
-}