+++ common/src/main/java/net/bettercombat/client/animation/IExtendedAnimation.java
@@ -1,5 +1,0 @@
-package net.bettercombat.client.animation;
-
-public interface IExtendedAnimation {
-    boolean isActiveInFirstPerson();
-}