+++ common/src/main/java/net/bettercombat/client/animation/PoseSubStack.java
@@ -7,6 +7,7 @@ import dev.kosmx.playerAnim.api.layered.modifier.AbstractModifier;
 import dev.kosmx.playerAnim.api.layered.modifier.MirrorModifier;
 import dev.kosmx.playerAnim.core.data.KeyframeAnimation;
 import dev.kosmx.playerAnim.core.util.Ease;
+import net.bettercombat.client.animation.first_person.CustomAnimationPlayer;
 
 import javax.annotation.Nullable;
 import java.util.function.Consumer;