+++ common/src/main/java/net/bettercombat/logic/knockback/ConfigurableKnockback.java
@@ -1,0 +1,5 @@
+package net.bettercombat.logic.knockback;
+
+public interface ConfigurableKnockback {
+    void setKnockbackMultiplier_BetterCombat(float strength);
+}