+++ common/src/main/java/net/bettercombat/mixin/client/firstpersonrender/EntityRenderDispatcherMixin.java
@@ -1,6 +1,6 @@
 package net.bettercombat.mixin.client.firstpersonrender;
 
-import net.bettercombat.client.animation.FirstPersonRenderHelper;
+import net.bettercombat.client.animation.first_person.FirstPersonRenderHelper;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.EntityRenderDispatcher;
 import net.minecraft.client.util.math.MatrixStack;