+++ common/src/main/java/net/bettercombat/network/ServerNetwork.java
@@ -10,7 +10,9 @@ import net.bettercombat.logic.PlayerAttackHelper;
 import net.bettercombat.logic.PlayerAttackProperties;
 import net.bettercombat.logic.TargetHelper;
 import net.bettercombat.logic.WeaponRegistry;
+import net.bettercombat.logic.knockback.ConfigurableKnockback;
 import net.bettercombat.mixin.LivingEntityAccessor;
+import net.bettercombat.utils.MathHelp;
 import net.bettercombat.utils.SoundHelper;
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
 import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
@@ -24,7 +26,6 @@ import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.decoration.ArmorStandEntity;
-import net.minecraft.entity.mob.HostileEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.network.packet.c2s.play.PlayerInteractEntityC2SPacket;
@@ -54,7 +55,7 @@ public class ServerNetwork {
                 return;
             }
             final var packet = Packets.AttackAnimation.read(buf);
-            final var forwardBuffer = new Packets.AttackAnimation(player.getId(), packet.isOffHand(), packet.animationName(), packet.length(), packet.upswing()).write();
+            final var forwardBuffer = new Packets.AttackAnimation(player.getId(), packet.animatedHand(), packet.animationName(), packet.length(), packet.upswing()).write();
             PlayerLookup.tracking(player).forEach(serverPlayer -> {
                 try {
                     if (serverPlayer.getId() != player.getId() && ServerPlayNetworking.canSend(serverPlayer, Packets.AttackAnimation.ID)) {
@@ -116,6 +117,10 @@ public class ServerNetwork {
                     SoundHelper.playSound(world, player, attack.swingSound());
                 }
 
+                var attackCooldown = PlayerAttackHelper.getAttackCooldownTicksCapped(player);
+                var knockbackMultiplier = BetterCombat.config.knockback_reduced_for_fast_attacks
+                        ? MathHelp.clamp(attackCooldown / 12.5F, 0.1F, 1F)
+                        : 1F;
                 var lastAttackedTicks = ((LivingEntityAccessor)player).getLastAttackedTicks();
                 if (!useVanillaPacket) {
                     player.setSneaking(request.isSneaking());
@@ -135,8 +140,13 @@ public class ServerNetwork {
                             || (entity instanceof ArmorStandEntity && ((ArmorStandEntity)entity).isMarker())) {
                         continue;
                     }
-                    if (BetterCombat.config.allow_fast_attacks && entity instanceof LivingEntity) {
-                        ((LivingEntity)entity).timeUntilRegen = 0;
+                    if (entity instanceof LivingEntity livingEntity) {
+                        if (BetterCombat.config.allow_fast_attacks) {
+                            livingEntity.timeUntilRegen = 0;
+                        }
+                        if (knockbackMultiplier != 1F) {
+                            ((ConfigurableKnockback)livingEntity).setKnockbackMultiplier_BetterCombat(knockbackMultiplier);
+                        }
                     }
                     ((LivingEntityAccessor) player).setLastAttackedTicks(lastAttackedTicks);
                     // System.out.println("Server - Attacking hand: " + (hand.isOffHand() ? "offhand" : "mainhand") + " CD: " + player.getAttackCooldownProgress(0));
@@ -153,6 +163,11 @@ public class ServerNetwork {
                                 return;
                             }
                             player.attack(entity);
+                        }
+                    }
+                    if (entity instanceof LivingEntity livingEntity) {
+                        if (knockbackMultiplier != 1F) {
+                            ((ConfigurableKnockback)livingEntity).setKnockbackMultiplier_BetterCombat(1F);
                         }
                     }
                 }