+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.5.5
+- Add attack speed cap
+- Add knockback reduction for fast attacks (configurable)
+- Improve back to back animation overlapping
+- Fix offhand arm invisible while 2H animations
+- Rework feinting
+- Update `staff` and `battlestaff` presets
+
 # 1.5.4
 - Fix attacking hostile mounts/vehicles (Thanks to Emily Darrow)
 - Fix cleaving for decoration items (such as Item Frames)